+++ .github/workflows/ci.yml
@@ -1,9 +1,10 @@
-name: Rust
+name: ci
 
 on:
   push:
     branches: [ "main" ]
   pull_request:
+    branches: [ "main" ]
     types:
       - opened
       - synchronize
@@ -12,10 +13,8 @@ env:
   CARGO_TERM_COLOR: always
 
 jobs:
-  build:
-
+  test:
     runs-on: ubuntu-latest
-
     steps:
     - uses: actions/checkout@v3
     - name: Build
